Enforce the usage caps, which have never been enforced
checkBudget and checkSkillBudget sum today's usage with
COALESCE(SUM(mtok_in + mtok_out), 0). Both columns are BIGINT, and
Postgres widens SUM(bigint) to numeric, which does not decode into the
Int64 the row type asks for. So every budget check threw a
PostgresDecodingError, every one landed in the fail-open catch, and
neither the trial token cap nor the per-skill cap has ever applied. The
fail-open has been there since HER-175.

It stayed invisible because the error is logged and swallowed, and
because no test ran the check against Postgres. The integration log
shows it plainly: 'usage_meter budget check failed' on every chat.

Fix is a ::BIGINT cast on both sums. UsageMeterBudgetTests runs allow,
degrade, deny and per-skill deny against a real database; three of the
four failed before the cast and pass after it.

This changes production behaviour when deployed. With the shipped
defaults, a trial tenant over 800k tokens a day is degraded to the
cheaper model, over 1M is refused until UTC midnight, and a skill over
200k a day is refused. That is the designed behaviour; it simply never
ran.

+++ Sources/App/Billing/UsageMeterService.swift
@@ -161,6 +161,10 @@ actor UsageMeterService {
             return .allow
         }
 
+        // `SUM` over BIGINT columns returns Postgres `numeric`, which does not
+        // decode into `Int64`. Without the `::BIGINT` cast every check threw,
+        // fell into the fail-open `catch` below, and the cap was never enforced.
+        // `UsageMeterBudgetTests` exercises this against a real database.
         struct UsageRow: Decodable {
             let total: Int64
 
@@ -171,7 +175,7 @@ actor UsageMeterService {
 
         do {
             let row = try await sql.raw("""
-            SELECT COALESCE(SUM(mtok_in + mtok_out), 0) AS total
+            SELECT COALESCE(SUM(mtok_in + mtok_out), 0)::BIGINT AS total
             FROM usage_meter
             WHERE tenant_id = \(bind: tenantID)
               AND day = CURRENT_DATE
@@ -228,7 +232,7 @@ actor UsageMeterService {
         let prefix = "skill:\(skillName)/"
         do {
             let row = try await sql.raw("""
-            SELECT COALESCE(SUM(mtok_in + mtok_out), 0) AS total
+            SELECT COALESCE(SUM(mtok_in + mtok_out), 0)::BIGINT AS total
             FROM usage_meter
             WHERE tenant_id = \(bind: tenantID)
               AND day = CURRENT_DATE
